Task 5: Minimal distance algorithm implemented

+++ husky_highlevel_controller/include/husky_highlevel_controller/Algorithm.hpp
@@ -6,7 +6,7 @@ namespace husky_highlevel_controller {
         public:
             Algorithm();
             virtual ~Algorithm();
-            float getMinimalDistance(const sensor_msgs::LaserScan& msg);
+            double getMinimalDistance(const sensor_msgs::LaserScan& msg);
         private:
     };
 } // end namespace
